requirements.txt
requests~=2.31.0
erniebot~=0.5.5
langchain~=0.2.0
pydantic~=2.7.1
python-dotenv~=1.0.1
numpy~=1.24.3